+++ CHANGELOG.md
@@ -9,6 +9,12 @@ NOTE: As semantic versioning states all 0.y.z releases can contain breaking chan
 
 We use *breaking* word for marking changes that are not backward compatible (relates only to v0.y.z releases.)
 
+## Unreleased
+
+### Fixed
+
+* Fixed case with installing latest binary version, when binary was not installed before.
+
 ## [v0.2.2](https://github.com/bwplotka/bingo/releases/tag/v0.2.2) - 2020.06.10
 
 ### Fixed